Initial create command.
