+++ src/Commands/Self/Plugin/PluginBaseCommand.php
@@ -7,6 +7,7 @@ use Pantheon\Terminus\Exceptions\TerminusNotFoundException;
 use Pantheon\Terminus\Helpers\LocalMachineHelper;
 use Pantheon\Terminus\Plugins\PluginDiscovery;
 use Pantheon\Terminus\Exceptions\TerminusException;
+use Pantheon\Terminus\Plugins\PluginInfo;
 
 /**
  * Class PluginBaseCommand
@@ -28,6 +29,10 @@ abstract class PluginBaseCommand extends TerminusCommand
         . DIRECTORY_SEPARATOR . "backup.tar.gz \"{dir}\"";
     const COMPOSER_REMOVE_REPOSITORY = 'composer config -d {dir} --unset repositories.{repo_name}';
     const DEPENDENCIES_UPDATE_COMMAND = 'composer update -d {dir} {packages} --with-dependencies';
+    const CONFIGURE_PATH_REPO_COMMAND =
+    'composer config -d {dir} repositories.{name} path {path}';
+    const INSTALL_COMMAND =
+    'composer require -d {dir} {project} --no-update';
 
     /**
      * @var array|null
@@ -330,5 +335,97 @@ abstract class PluginBaseCommand extends TerminusCommand
     protected function ensureDirectoryExists($path, $permissions = 0755)
     {
         $this->getLocalMachine()->getFileSystem()->mkdir($path, $permissions);
+    }
+
+    /**
+     * Gets project name from given path.
+     */
+    protected function getProjectNameFromPath($project_or_path)
+    {
+        $composerJson = $project_or_path . '/composer.json';
+        $composerContents = file_get_contents($composerJson);
+        // If the specified dir does not contain a terminus plugin, throw an error
+        $composerData = json_decode($composerContents, true);
+        if (!isset($composerData['type']) || ($composerData['type'] !== 'terminus-plugin')) {
+            throw new TerminusException(
+                'Cannot install from path {path} because the project there is not of type "terminus-plugin"',
+                ['path' => $project_or_path]
+            );
+        }
+
+        // If the specified dir does not have a name in the composer.json, throw an error
+        if (empty($composerData['name'])) {
+            throw new TerminusException(
+                'Cannot install from path {path} because the project there does not have a name',
+                ['path' => $project_or_path]
+            );
+        }
+
+        // Finally, return the project name and let install command install it as normal.
+        return $composerData['name'];
+    }
+
+    /**
+     * @param string $project_name Name of project to be installed
+     * @param string $instalation_path If not empty, install as a path repository
+     * @return array Results from the install command
+     */
+    protected function installProject($project_name, $instalation_path = '')
+    {
+        $plugin_name = PluginInfo::getPluginNameFromProjectName($project_name);
+        $project_name_parts = explode(':', $project_name);
+        $project_name_without_version = reset($project_name_parts);
+        $config = $this->getConfig();
+        $original_plugins_dir = $config->get('plugins_dir');
+        $original_dependencies_dir = $config->get('terminus_dependencies_dir');
+        $folders = $this->updateTerminusDependencies($original_plugins_dir, $original_dependencies_dir);
+        $plugins_dir = $folders['plugins_dir'];
+        $dependencies_dir = $folders['dependencies_dir'];
+        try {
+            if (!empty($instalation_path)) {
+                // Update path repository in plugins dir and dependencies dir.
+                foreach ([$plugins_dir, $dependencies_dir] as $dir) {
+                    $command = str_replace(
+                        ['{dir}', '{name}', '{path}',],
+                        [$dir, $project_name_without_version, realpath($instalation_path),],
+                        self::CONFIGURE_PATH_REPO_COMMAND
+                    );
+                    $results = $this->runCommand($command);
+                    if ($results['exit_code'] !== 0) {
+                        throw new TerminusException(
+                            'Error configuring path repository in ' . basename($dir),
+                            []
+                        );
+                    }
+                }
+            }
+
+            $command = str_replace(
+                ['{dir}', '{project}',],
+                [$plugins_dir, $project_name,],
+                self::INSTALL_COMMAND
+            );
+            $results = $this->runCommand($command);
+            if ($results['exit_code'] !== 0) {
+                throw new TerminusException(
+                    'Error requiring package in terminus-plugins.',
+                    []
+                );
+            }
+            $results = $this->runComposerUpdate($dependencies_dir, $project_name_without_version);
+            if ($results['exit_code'] !== 0) {
+                throw new TerminusException(
+                    'Error running composer update in terminus-dependencies.',
+                    []
+                );
+            }
+            $this->replaceFolder($plugins_dir, $original_plugins_dir);
+            $this->replaceFolder($dependencies_dir, $original_dependencies_dir);
+            $this->log()->notice('Installed {project_name}.', compact('project_name'));
+        } catch (TerminusException $e) {
+            $this->log()->error($e->getMessage());
+        }
+
+        return $results;
     }
 }
